[rtl] move Bundles in V to Bundles.scala
