+++ v/src/V.scala
@@ -31,51 +31,6 @@ case class VParam(XLEN: Int = 32, dataPathWidth: Int = 32, VLEN: Int = 1024, lan
   require(XLEN == dataPathWidth)
 }
 
-class VReq(param: VParam) extends Bundle {
-  val inst:     UInt = UInt(32.W)
-  val src1Data: UInt = UInt(param.XLEN.W)
-  val src2Data: UInt = UInt(param.XLEN.W)
-}
-
-class VResp(param: VParam) extends Bundle {
-  // todo: vector解出来是否需要写rd？
-  val data: UInt = UInt(param.XLEN.W)
-}
-
-class InstRecord(param: VParam) extends Bundle {
-  val instIndex: UInt = UInt(param.instIndexSize.W)
-  val vrfWrite:  Bool = Bool()
-
-  /** Whether operation is widen */
-  val w: Bool = Bool()
-
-  /** Whether operation is narrowing */
-  val n: Bool = Bool()
-  // load | store
-  val ls: Bool = Bool()
-}
-
-class InstState extends Bundle {
-  val wLast:    Bool = Bool()
-  val idle:     Bool = Bool()
-  val sExecute: Bool = Bool()
-  val sCommit:  Bool = Bool()
-}
-
-class specialInstructionType extends Bundle {
-  val red:      Bool = Bool()
-  val compress: Bool = Bool()
-  val viota:    Bool = Bool()
-  // 其他的需要对齐的指令
-  val other: Bool = Bool()
-}
-
-class InstControl(param: VParam) extends Bundle {
-  val record: InstRecord = new InstRecord(param)
-  val state:  InstState = new InstState
-  val endTag: Vec[Bool] = Vec(param.lane + 1, Bool())
-}
-
 class V(val parameter: VParam) extends Module with SerializableModule[VParam] {
   val req:              DecoupledIO[VReq] = IO(Flipped(Decoupled(new VReq(parameter))))
   val resp:             ValidIO[VResp] = IO(Valid(new VResp(parameter)))
